app: add post component
